requirements.txt
j2cli==0.3.10
python-slugify==8.0.4
